+++ src/entities/brand-protection/seller-profile/model/types.ts
@@ -1,0 +1,56 @@
+export enum SellerProfileCategory {
+  RISKY = 1,
+  SECOND = 2,
+  PARALLEL = 3,
+  IGNORED = 4,
+  THIRD = 5,
+  NEW = 6,
+  BRANDABUSE = 7,
+  OFFICIAL = 8,
+  REPORTED = 9,
+}
+
+export interface SellerProfile {
+  id: number;
+  name: string;
+  brands: Array<{ name: string }>;
+
+  address: string;
+  email: string;
+  mersisNumber: string;
+  phoneNumber: string;
+  taxNumber: string;
+
+  // Insights
+  isClosed: boolean;
+  category: SellerProfileCategory;
+  platformCount: number;
+
+  platforms: Array<{ id: number; name: string; iconLink: string }>;
+  sellers: Array<{
+    id: number;
+    name: string;
+    avatarUrl: string;
+    rating: number;
+    url: string;
+    isClosed: boolean;
+    platform: {
+      id: number;
+      name: string;
+      iconLink: string;
+    };
+  }>;
+  stats: {
+    riskyProductCount: {
+      [brandId: string]: number;
+    };
+    closedProductCount: {
+      [brandId: string]: number;
+    };
+    productCountPlatformDispersion: {
+      [brandId: string]: {
+        [platform: string]: number;
+      };
+    };
+  };
+}