+++ src/pages/seller-profile-detail/model/index.ts
@@ -1,0 +1,1 @@
+export { useSuspenseSellerProfileData } from "./hooks";