+++ src/shared/api/brand-protection/bp-api.service.ts
@@ -1,3 +1,5 @@
+import { AppError } from "@/shared/lib/error-handler";
+
 import { bpApi } from "./bp-api.instance";
 import { BP_API_ENDPOINTS } from "./bp-api.config";
 import {
@@ -8,6 +10,12 @@ import {
   GroupBrandsResponseDto,
   GroupBrandsQueryDto,
   UpdateSelectedCompanyRequestDto,
+  GetProfileByIdParams,
+  GetProfileByIdResponse,
+  ProductsQueryDto,
+  ProductsResponseDto,
+  ProductDto,
+  UpdateProfileRequestDto,
 } from "./bp-api.types";
 import {
   RefreshTokenRequestDtoSchema,
@@ -17,6 +25,12 @@ import {
   GroupBrandsQueryDtoSchema,
   GroupBrandsResponseDtoSchema,
   UpdateSelectedCompanyRequestDtoSchema,
+  GetProfileByIdParamsSchema,
+  GetProfileByIdResponseSchema,
+  ProductsQueryDtoSchema,
+  ProductsResponseDtoSchema,
+  ProductResponseDtoSchema,
+  UpdateProfileRequestDtoSchema,
 } from "./bp-api.schemas";
 
 // --------------------------
@@ -65,4 +79,90 @@ export const getGroupBrands = (params: GroupBrandsQueryDto) => {
       responseData: GroupBrandsResponseDtoSchema,
     },
   });
+};
+
+// --------------------------
+// Profile Services => /profile
+// --------------------------
+export const getProfileById = async (
+  id: number,
+  params: GetProfileByIdParams
+) => {
+  const response = await bpApi.get<GetProfileByIdResponse>(
+    `${BP_API_ENDPOINTS.profiles}/${id}`,
+    {
+      params,
+      validationSchemas: {
+        params: GetProfileByIdParamsSchema,
+        responseData: GetProfileByIdResponseSchema,
+      },
+      headers: {
+        "Cache-Control": "no-cache",
+      },
+    }
+  );
+
+  if (!response.success) throw new AppError(response.error);
+
+  return response.data;
+};
+
+export const updateProfile = async (
+  id: number,
+  data: UpdateProfileRequestDto
+) => {
+  const response = await bpApi.patch(
+    `${BP_API_ENDPOINTS.profiles}/${id}`,
+    data,
+    {
+      validationSchemas: {
+        data: UpdateProfileRequestDtoSchema,
+      },
+    }
+  );
+
+  if (!response.success) throw new AppError(response.error);
+
+  return response.data;
+};
+
+// --------------------------
+// Products Services => /products
+// --------------------------
+export const getProductById = async (id: string) => {
+  const response = await bpApi.get<ProductDto>(
+    `${BP_API_ENDPOINTS.products}/${id}`,
+    {
+      validationSchemas: {
+        responseData: ProductResponseDtoSchema,
+      },
+      headers: {
+        "Cache-Control": "no-cache",
+      },
+    }
+  );
+
+  if (!response.success) throw new AppError(response.error);
+
+  return response.data;
+};
+
+export const getProducts = async (params: ProductsQueryDto) => {
+  const response = await bpApi.get<ProductsResponseDto>(
+    BP_API_ENDPOINTS.products,
+    {
+      params,
+      validationSchemas: {
+        params: ProductsQueryDtoSchema,
+        responseData: ProductsResponseDtoSchema,
+      },
+      headers: {
+        "Cache-Control": "no-cache",
+      },
+    }
+  );
+
+  if (!response.success) throw new AppError(response.error);
+
+  return response.data;
 };